ReferenceSpeech/scripts/extract_features/extract_linguistic/requirements/wlcoref_pip_spec_file.txt
attrs==22.2.0
backcall==0.2.0
charset-normalizer==3.0.1
click==8.1.3
decorator==5.1.1
emoji==2.2.0
filelock==3.9.0
idna==3.4
importlib-metadata==6.0.0
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
jsonlines==3.1.0
matplotlib-inline==0.1.6
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.36
protobuf==4.21.12
ptyprocess==0.7.0
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sacremoses==0.0.53
sentencepiece==0.1.97
six==1.16.0
stanza==1.4.2
tokenizers==0.8.1rc2
toml==0.10.2
torch==1.4.0
torchvision==0.5.0
tqdm==4.64.1
traitlets==5.9.0
transformers==3.2.0
typing_extensions==4.4.0
urllib3==1.26.14
wcwidth==0.2.6
zipp==3.12.1
